Factor out LandlordMailingAddress component.
